+++ src/utils/getYAxisDomain.ts
@@ -1,0 +1,11 @@
+// Tremor Raw getYAxisDomain [v0.0.0]
+
+export const getYAxisDomain = (
+  autoMinValue: boolean,
+  minValue: number | undefined,
+  maxValue: number | undefined,
+) => {
+  const minDomain = autoMinValue ? "auto" : minValue ?? 0
+  const maxDomain = maxValue ?? "auto"
+  return [minDomain, maxDomain]
+}