+++ src/utils/changelog.md
@@ -1,7 +1,15 @@
 # Tremor Raw utils Changelog
 
+## 0.1.0
+
+### Changes
+
+- New: chartColors utilities
+- New: getYAxisDomain utility
+- New: hasOnlyOneValueForKey utility
+
 ## 0.0.1
 
 ### Changes
 
-* Updated sort order focusInput, focusRing, hasErrorInput
+- Updated sort order focusInput, focusRing, hasErrorInput